Type checker improvements, create distinct symbols
Type checker checks do not report related errors (e.g. don't also report that the if condition is non-boolean when the true error is that it references to an unknown VarDecl)

There are now two types of Symbols, which don't inherit code of each other (because of the advantage of type safety when then gain).